requirements.txt
chardet==5.2.0
openai==1.60.2
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
